requirements.txt
Django==5.2.5
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.7.0
pytest==8.0.0
pytest-playwright==0.4.3
pytest-html==4.1.1